+++ SleepyEngine/GUI/GUIElement.cpp
@@ -1,5 +1,5 @@
 #include "GUIElement.h"
 
-GUIElement::GUIElement( u32 x, u32 y, u32 width, u32 height )
+GUIElement::GUIElement( f32 x, f32 y, f32 width, f32 height )
 	: m_iX( x ), m_iY( y), m_iWidth( width ), m_iHeight( height )
 {}
